Add logical local-timestamp types
Added the two remaining logical types from [the Avro specification][1]:

  * `local-timestamp-millis`
  * `local-timestamp-micros`

The encoding for these types is the same as `timestamp-millis` and `timestamp-micros` respectively, but they should be interpreted as *local* timestamps rather than UTC. This corresponds to the [`LocalTime`][2] type from the Haskell [time][3] package.

I also wrote Haddock documentation for all of the logical types supported by this package.

[1]: https://avro.apache.org/docs/current/spec.html#Logical+Types

[2]: https://hackage.haskell.org/package/time-1.13/docs/Data-Time-LocalTime.html#t:LocalTime

[3]: https://hackage.haskell.org/package/time